Updated Android Project #1
This update incorporates the user login / registration. Work has also
been done towards the search function - with the extra time allocated,
it should be possible to create code that resolves a GPS co-ordinate to
a location somehow. The remember me function broke after adding the
AsyncTasks that process the registration / login (this will be fixed).
The update profile function is working.