Latest Update (One More Incoming)
This pull request will serve as a backup of the latest working version
of the mobile app. In this update, I have extended the
"EventDetailsFragment" and added the mechanism that registers the user
to the Event (there is a glitch there I am aware of). In the next update
I will add the view that shows all registered events.  I also added the
new PHP scripts used.